Refactor: inline MTP device state and verification into decode_fwd_mtp
decode_device_state and decode_mtp_verify were imported by decode_fwd_mtp
and nothing else, and their three kernels are consecutive stages of one
state machine: prepare loads a request's previous tail/draft, verify packs
the committed window, advance commits the result back to the same slot.
Splitting them across two modules meant decode_fwd_mtp forwarded eight
STATE_* constants just to describe the pool it owns.

- Move the STATE_* layout, prepare_decode_from_device_state,
  advance_decode_device_state and verify_and_pack_mtp_tokens into
  decode_fwd_mtp.py and drop the two old modules
- Drop the standalone decode_mtp_verify test entry, golden and fixtures
- Bind S from config: decode_fwd re-exports B and T but not the per-step
  token count
- Repoint the model doc at the inlined stages

Kernel behavior is unchanged; decode_fwd_mtp still compiles.
